% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floyd_custom_xlri_function.R
\name{floyd_custom_xlri}
\alias{floyd_custom_xlri}
\title{A Floyd Function}
\usage{
floyd_custom_xlri()
}
\description{
This function allows you to solve network problems using floyd algorithm.
The function takes input in form of distance, for larger distance input veryhighvalue ~ 1000 times
}
